+++ lib/waha/rails.rb
@@ -4,5 +4,10 @@ require "waha"
 require "active_support"
 require "rails/railtie"
 require_relative "rails/controller_concern"
+require_relative "rails/session_state"
 require_relative "rails/railtie"
 require_relative "../generators/waha/install/install_generator"
+
+module Waha
+  SessionState = Waha::Rails::SessionState
+end